check if .env already exists and stop overwrite if so

+++ install.sh
@@ -21,8 +21,12 @@ reMoveMagentoEnv() {
 }
 
 createEnv() {
-    echo "cp ./.env.template ./.env";
-    cp ./.env.template ./.env
+    if [ ! -f ./.env ]; then
+        echo "cp ./.env.template ./.env";
+        cp ./.env.template ./.env
+    else
+        echo ".env File exists already";
+    fi
 }
 
 dockerRefresh() {
@@ -123,7 +127,7 @@ install() {
         --use-rewrites=1;
 }
 
-setDomain() {
+setDomainAndCookieName() {
     SET_URL_SECURE="USE $1; INSERT INTO core_config_data(scope, value, path) VALUES('default', 'http://$5/', 'web/unsecure/base_url') ON DUPLICATE KEY UPDATE value='http://$5/', path='web/unsecure/base_url', scope='default';";
     SET_URL_UNSECURE="USE $1; INSERT INTO core_config_data(scope, value, path) VALUES('default', 'https://$5/', 'web/secure/base_url') ON DUPLICATE KEY UPDATE value='https://$5/', path='web/secure/base_url', scope='default';";
     SET_URL_COOKIE="USE $1; INSERT core_config_data(scope, value, path) VALUES('default', '$5', 'web/cookie/cookie_domain') ON DUPLICATE KEY UPDATE value='$5', path='web/cookie/cookie_domain', scope='default';";
@@ -202,7 +206,7 @@ reMoveMagentoEnv
 dockerRefresh
 composerPackages ${USER} ${NAMESPACE}_php_${PHP_VERSION_SET} ${SHOP_URI}
 install ${USER} ${SHOP_URI} ${NAMESPACE}_php_${PHP_VERSION_SET} ${NAMESPACE} ${MYSQL_USER} ${MYSQL_PASSWORD} ${SSL}
-setDomain ${NAMESPACE} ${MYSQL_USER} ${MYSQL_PASSWORD} ${NAMESPACE}_db ${SHOP_URI}
+setDomainAndCookieName ${NAMESPACE} ${MYSQL_USER} ${MYSQL_PASSWORD} ${NAMESPACE}_db ${SHOP_URI}
 exchangeMagentoEnv ${USER} ${NAMESPACE}_nginx
 magentoRefresh ${USER} ${NAMESPACE}_php_${PHP_VERSION_SET} ${SHOP_URI}
 getMagerun ${SHOP_URI}
